Dynamically position the categories dropdown on desktop
