+++ modules/Header/ui/CategoriesNav/CategoriesNavDesktop.tsx
@@ -17,11 +17,12 @@ import { FeaturedCategory } from './FeaturedCategory';
 import styles from './CategoriesNavDesktop.module.scss';
 
 export function CategoriesNavDesktop({
-    categories,
-    translatedCategories,
-    localeCode,
     buttonClassName,
+    categories,
+    localeCode,
+    marginTop,
     navigationItemClassName,
+    translatedCategories,
 }: CategoriesNavDesktop.Props) {
     function getCategory(translated: TranslatedCategory) {
         return categories.find((category) => category.id === translated.id)!;
@@ -63,7 +64,7 @@ export function CategoriesNavDesktop({
                             leaveTo={styles.transitionOpenStart}
                         >
                             <div>
-                                <Popover.Panel className={styles.popover}>
+                                <Popover.Panel className={styles.popover} style={{ marginTop }}>
                                     <div className={styles.container}>
                                         <div
                                             className={classNames(styles.grid, {
@@ -124,7 +125,7 @@ export function CategoriesNavDesktop({
                                         </div>
                                     </div>
                                 </Popover.Panel>
-                                <div className={styles.backdrop} />
+                                <div className={styles.backdrop} style={{ marginTop }} />
                             </div>
                         </Transition>
                         {open && <BlockPageScroll />}
@@ -147,10 +148,11 @@ function BlockPageScroll() {
 
 export namespace CategoriesNavDesktop {
     export interface Props {
-        categories: Category[];
-        translatedCategories: TranslatedCategory[];
-        localeCode: Locale.Code;
         buttonClassName?: string;
+        categories: Category[];
+        localeCode: Locale.Code;
+        marginTop: number | undefined;
         navigationItemClassName?: string;
+        translatedCategories: TranslatedCategory[];
     }
 }
